Fix boss fight health bar

+++ src/main/kotlin/character/BossFight.kt
@@ -28,7 +28,7 @@ class BossFight(
     }
 
     fun tick() {
-        bossBar.progress((character.health / character.maxHealth).toFloat())
+        bossBar.progress((character.health.toFloat() / character.maxHealth.toFloat()))
 
         viewers.filter {
             it.removed || Position.sqrDistance(
